pyproject.toml
[project]
name = "dnstwist-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dnstwist[full]>=20240812",
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "google-cloud-pubsub>=2.28.0",
    "mysql-connector-python>=9.2.0",
    "mysqlclient>=2.2.7",
    "pip>=24.3.1",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
    "tenacity>=9.0.0",
]
